Moved cube_to_dp and cubemap logic into CopyEffects
