+++ servers/rendering/renderer_rd/shaders/effects/cube_to_dp.glsl
