+++ servers/rendering/renderer_rd/shaders/effects/cubemap_downsampler_raster.glsl
