+++ servers/rendering/renderer_rd/shaders/effects/cubemap_filter_raster.glsl
