+++ servers/rendering/renderer_rd/shaders/effects/cubemap_roughness.glsl
