+++ servers/rendering/renderer_rd/shaders/effects/cubemap_roughness_raster.glsl
